requirements.txt
fastapi[standard]>=0.111.0
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
celery>=5.4.0
redis>=5.0.4
anthropic>=0.26.0
google-api-python-client>=2.129.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
cryptography>=42.0.7
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.2.1
requests>=2.32.0
httpx>=0.27.0
pytest>=8.2.1
pytest-asyncio>=0.23.6
google-genai==1.47.0
